+++ packages/genesys-spark-components/src/components/stable/gux-datepicker/gux-datepicker.tsx
@@ -25,7 +25,9 @@ import { OnClickOutside } from '@utils/decorator/on-click-outside';
 import { trackComponent } from '@utils/tracking/usage';
 import { CalendarModes } from '../../../common-enums';
 import { buildI18nForComponent, GetI18nValue } from '../../../i18n';
-
+import * as sparkIntl from '../../../genesys-spark-utils/intl';
+// Remove with this ticket https://inindca.atlassian.net/browse/COMUI-2598
+import { readRegionalDatesCookie } from '../../../i18n/check-regional-dates-cookie';
 import { GuxCalendarDayOfWeek } from '../gux-calendar/gux-calendar.types';
 
 import translationResources from './i18n/en.json';
@@ -122,8 +124,8 @@ export class GuxDatepicker {
   /**
    * The datepicker date format (default to mm/dd/yyyy, or specified)
    */
-  @Prop()
-  format: string = 'mm/dd/yyyy';
+  @Prop({ mutable: true })
+  format: string;
 
   /**
    * Disable the input and prevent interactions.
@@ -157,6 +159,12 @@ export class GuxDatepicker {
    */
   @State()
   active: boolean = false;
+
+  /**
+   * Tracks the amount of key presses when focusing on the input.
+   */
+  @State()
+  pressedKeys: Array<string> = [];
 
   @Watch('value')
   watchValue() {
@@ -240,6 +248,8 @@ export class GuxDatepicker {
           );
           this.setIntervalRange(previousIntervalRange);
           this.setCursorRange();
+          this.pressedKeys.length = 0;
+
           break;
         }
         case 'ArrowRight': {
@@ -251,6 +261,8 @@ export class GuxDatepicker {
           );
           this.setIntervalRange(nextIntervalRange);
           this.setCursorRange();
+          this.pressedKeys.length = 0;
+
           break;
         }
         default:
@@ -312,6 +324,7 @@ export class GuxDatepicker {
   @OnClickOutside({ triggerEvents: 'mousedown' })
   onClickOutside() {
     this.active = false;
+    this.pressedKeys.length = 0;
   }
 
   /*********  Event Handlers  **********/
@@ -362,6 +375,7 @@ export class GuxDatepicker {
   }
 
   updateIntervalValue(event: KeyboardEvent): void {
+    this.pressedKeys.push(event.key);
     const inputNumber = parseInt(event.key, 10);
 
     if (!isNaN(inputNumber)) {
@@ -377,7 +391,7 @@ export class GuxDatepicker {
       ) {
         this.typeYearValue(currentSectionValue, event.key);
       } else {
-        if (this.canSetDate(inputNumber)) {
+        if (this.canSetDate(inputNumber) && this.pressedKeys.length >= 2) {
           this.updateSelection(
             this.focusedField,
             `${currentSectionValue[1]}${event.key}`
@@ -479,7 +493,7 @@ export class GuxDatepicker {
     }
   }
 
-  canSetDate(key: number) {
+  canSetDate(key: number): boolean {
     const newValue = parseInt(
       [
         this.focusedField.value[this.intervalRange.selectionEnd - 1].toString(),
@@ -645,7 +659,13 @@ export class GuxDatepicker {
   async componentWillLoad() {
     trackComponent(this.root, { variant: this.mode });
     this.i18n = await buildI18nForComponent(this.root, translationResources);
-
+    if (readRegionalDatesCookie()) {
+      this.format =
+        this.format ||
+        sparkIntl.getFormat(sparkIntl.determineDisplayLocale(this.root));
+    } else {
+      this.format = this.format || 'mm/dd/yyyy';
+    }
     this.watchMinDate(this.minDate);
     this.watchMaxDate(this.maxDate);
     this.watchFormat(this.format);