+++ packages/genesys-spark-components/src/components/stable/gux-listbox/gux-listbox.tsx
@@ -175,8 +175,8 @@ export class GuxListbox {
     );
   }
 
-  // eslint-disable-next-line @typescript-eslint/require-await
   @Method()
+  // eslint-disable-next-line @typescript-eslint/require-await
   async guxSelectActive(): Promise<void> {
     actOnActiveOption(this.root, value => this.updateValue(value));
   }