+++ packages/genesys-spark-components/src/components/stable/gux-tabs-advanced/gux-tab-advanced-list/gux-tab-advanced-list.tsx
@@ -344,8 +344,8 @@ export class GuxTabAdvancedList {
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/require-await
   @Method()
+  // eslint-disable-next-line @typescript-eslint/require-await
   async guxSetActive(activeTab: string): Promise<void> {
     this.tabTriggers.forEach((tabTrigger, index) => {
       const active = tabTrigger.tabId === activeTab;