+++ packages/genesys-spark-components/src/i18n/check-regional-dates-cookie.ts
@@ -1,3 +1,4 @@
+// Remove with this ticket https://inindca.atlassian.net/browse/COMUI-2598
 export const readRegionalDatesCookie = () => {
   return !!document.cookie
     .split('; ')