+++ shared-configs/eslint-config-genesys-spark-components/typescript.js
@@ -7,6 +7,7 @@ module.exports = {
   ],
   parser: '@typescript-eslint/parser',
   rules: {
+    curly: 'error',
     '@typescript-eslint/no-inferrable-types': 'off',
     '@typescript-eslint/no-redundant-type-constituents': 'off',
     '@typescript-eslint/no-unsafe-enum-comparison': 'off'